requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==6.0.3
django-import-export==4.4.0
django-jazzmin==3.0.4
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.5
tablib==3.9.0
